ejercicio 4.21 completado

+++ controllers/blogs.js
@@ -13,10 +13,10 @@ blogsRouter.get("/", async (req, res) => {
 blogsRouter.post("/", async (req, res) => {
   const body = req.body;
 
-  const decodedToken = jwt.verify(req.token, process.env.SECRET)
-  
-  if(!decodedToken.id){
-    return res.status(401).json({error: 'token invalid'})
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" });
   }
 
   const user = await User.findById(decodedToken.id);
@@ -38,8 +38,19 @@ blogsRouter.post("/", async (req, res) => {
 
 // delete
 blogsRouter.delete("/:id", async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id);
-  res.status(204).end();
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  const blog = await Blog.findById(req.params.id);
+  const userBlog = blog.user;
+
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" });
+  } else if (userBlog.toString() === decodedToken.id) { //user verify
+    await Blog.findByIdAndDelete(req.params.id);
+    res.status(204).end();
+  } else {
+    return res.status(401).json({ error: "invalid user" });
+  }
 });
 
 // update
